+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Remove the requirement that all `Subscriber` methods, `Stream.subscribe(...)` and `Subscription.cancel()`
   method be invoked within the context of a VPU. This adds some unnecessary overhead and code complexity with
   no benefit when run on the target platform (i.e. Javascript).
+* Expose the `VirtualProcessorUnit.queue(Runnable)` method to enable queuing tasks directly on the VPU.
 
 ### [v0.04](https://github.com/spritz/spritz/tree/v0.04) (2019-03-12)
 [Full Changelog](https://github.com/spritz/spritz/compare/v0.03...v0.04)